+++ packages/editor/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @hexx/editor
 
+## 0.2.1
+
+### Patch Changes
+
+- inline marker works now
+
 ## 0.2.0
 
 ### Minor Changes